Delingstall skal være double, ikke integer (#1331)

+++ pensjon-brevbaker-api-model/src/main/kotlin/no/nav/pensjon/brev/api/model/maler/alderApi/AvslagUttakFoerNormertPensjonsalderAutoDto.kt
@@ -25,8 +25,8 @@ data class OpplysningerBruktIBeregningen(
     val uttaksgrad: Int,
     val trygdetid: Int,
     val pensjonsbeholdning: Kroner,
-    val delingstallVedUttak: Int,
-    val delingstallVedNormertPensjonsalder: Int?,
+    val delingstallVedUttak: Double,
+    val delingstallVedNormertPensjonsalder: Double?,
     val normertPensjonsalder: NormertPensjonsalder,
     val sisteOpptjeningsAar: Int?,
 )
